--- src/API/auth.js
@@ -1,55 +1,0 @@
-import apiUrl from '../apiConfig'
-import axios from 'axios'
-
-export const signUp = (credentials) => {
-  return axios({
-    method: 'POST',
-    url: apiUrl + '/sign-up/',
-    data: {
-      credentials: {
-        email: credentials.email,
-        password: credentials.password,
-        passwordConfirmation: credentials.passwordConfirmation
-      }
-    }
-  })
-}
-
-export const signIn = (credentials) => {
-  return axios({
-    url: apiUrl + '/sign-in/',
-    method: 'POST',
-    data: {
-      credentials: {
-        email: credentials.email,
-        password: credentials.password
-      }
-    }
-  })
-}
-
-export const signOut = (user) => {
-  return axios({
-    url: apiUrl + '/sign-out/',
-    method: 'DELETE',
-    headers: {
-      Authorization: `Bearer ${user.token}`
-    }
-  })
-}
-
-export const changePassword = (passwords, user) => {
-  return axios({
-    url: apiUrl + '/change-password/',
-    method: 'PATCH',
-    headers: {
-      Authorization: `Bearer ${user.token}`
-    },
-    data: {
-      passwords: {
-        old: passwords.oldPassword,
-        new: passwords.newPassword
-      }
-    }
-  })
-}
