--- src/components/auth/ChangePassword.js
@@ -1,39 +1,0 @@
-import React, { } from 'react'
-
-import Form from 'react-bootstrap/Form'
-import Button from 'react-bootstrap/Button'
-
-const ChangePassword = () => {
-  return (
-    <div className='row'>
-      <div className='col-sm-10 col-md-8 mx-auto mt-5'>
-        <h3>Change Password</h3>
-        <Form onSubmit={this.onChangePassword}>
-          <Form.Group controlId='oldPassword'>
-            <Form.Label>Old password</Form.Label>
-            <Form.Control
-              required
-              name='oldPassword'
-              type='password'
-              placeholder='Old Password'
-              onChange={this.handleChange}
-            />
-          </Form.Group>
-          <Form.Group controlId='newPassword'>
-            <Form.Label>New Password</Form.Label>
-            <Form.Control
-              required
-              name='newPassword'
-              type='password'
-              placeholder='New Password'
-              onChange={this.handleChange}
-            />
-          </Form.Group>
-          <Button variant='primary' type='submit'>Submit</Button>
-        </Form>
-      </div>
-    </div>
-  )
-}
-
-export default ChangePassword
