--- src/components/auth/SignOut.js
@@ -1,5 +1,0 @@
-const SignOut = () => {
-
-}
-
-export default SignOut
