--- src/components/header/header.scss
@@ -1,25 +1,0 @@
-#navbar {
-    .header {
-        width: 100%;
-        color:  rgb(1, 8, 65);
-   
-
-        background: rgba(1, 8, 65, 1);
-        background: linear-gradient(
-            180deg,
-            rgba(1, 8, 65, 1),
-            rgba(255, 255, 255, 1)
-        );
-
-        font-size: 2.5rem;
-        font-weight: 800;
-
-        padding: 1.5rem;
-        letter-spacing: 2px;
-        font-style: italic;
-
-        font-family: system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI",
-            Roboto, Oxygen, Ubuntu, Cantarell, "Open Sans", "Helvetica Neue",
-            sans-serif;
-    }
-}
