--- src/components/utils/Board.js
@@ -1,5 +1,0 @@
-/* eslint-disable */
-
-export class Board {
-    
-}
