+++ src/comps/settings/settings.scss
