+++ src/comps/table/table3/table3.scss
