+++ src/comps/utils/practice.js
@@ -1,0 +1,10 @@
+/* eslint-disable */
+
+class Sample {
+
+    static printMe() {
+        console.log("ME")
+    }
+}
+
+Sample.printMe()
